+++ test/Braintree.Tests.Integration/TransactionTransferTypeIntegrationTest.cs
@@ -53,7 +53,8 @@ namespace Braintree.Tests.Integration
                         LastName = "Silva",
                         MiddleName = "A",
                         DateOfBirth = new DateTime(2009, 1, 1),
-                        AccountReferenceNumber = "1000012345",
+                        AccountReferenceNumber = "987654321",
+                        AccountReferenceNumberType = "PHONE_NUMBER",
                         Address = new AddressRequest
                         {
                             StreetAddress = "1st Main, door 12th",
@@ -67,6 +68,8 @@ namespace Braintree.Tests.Integration
                         FirstName = "Bob",
                         LastName = "Souza",
                         MiddleName = "A",
+                        AccountReferenceNumber = "123456789",
+                        AccountReferenceNumberType = "BIC_SWIFT_CODE",
                         Address = new AddressRequest
                         {
                             StreetAddress = "1st Main, door 12th",
@@ -92,7 +95,7 @@ namespace Braintree.Tests.Integration
             var transactionRequest = new TransactionRequest
             {
                 Amount = 100.00M,
-                MerchantAccountId = "aft_first_data_wallet_transfer",
+                MerchantAccountId = MerchantAccountIDs.AFT_FIRST_DATA_WALLET_TRANSFER,
                 CreditCard = new TransactionCreditCardRequest
                 {
                     Number = "4111111111111111",
@@ -107,6 +110,86 @@ namespace Braintree.Tests.Integration
 
             Result<Transaction> result = gateway.Transaction.Sale(transactionRequest);
             Assert.IsFalse(result.IsSuccess());
+        }
+
+        [Test]
+        public void Sale_ShouldNotCreateTransactionWithInvalidSenderAccountReferenceNumberType()
+        {
+
+            var transactionRequest = new TransactionRequest
+            {
+                Amount = 100.00M,
+                MerchantAccountId = MerchantAccountIDs.AFT_FIRST_DATA_WALLET_TRANSFER,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = "4111111111111111",
+                    ExpirationDate = "06/2026",
+                    CVV = "123"
+                },
+                Transfer = new TransferRequest
+                {
+                    Type = "wallet_transfer",
+                    Sender = new SenderRequest
+                    {
+                        FirstName = "Bob",
+                        LastName = "Souza",
+                        MiddleName = "A",
+                        AccountReferenceNumber = "123456789",
+                        AccountReferenceNumberType = "INVALID_ACCOUNT_REFERENCE_NUMBER_TYPE",
+                        Address = new AddressRequest
+                        {
+                            StreetAddress = "1st Main, door 12th",
+                            Locality = "LA",
+                            Region = "CA",
+                            CountryCodeAlpha2 = "US"
+                        }
+                    }
+                }
+            };
+
+            Result<Transaction> result = gateway.Transaction.Sale(transactionRequest);
+            Assert.IsFalse(result.IsSuccess());
+            Assert.AreEqual(ValidationErrorCode.TRANSACTION_TRANSFER_SENDER_ACCOUNT_REFERENCE_NUMBER_TYPE_IS_NOT_VALID, result.Errors.ForObject("accountFundingTransaction").OnField("sender_account_reference_number_type")[0].Code);
+        }
+
+        [Test]
+        public void Sale_ShouldNotCreateTransactionWithInvalidReceiverAccountReferenceNumberType()
+        {
+
+            var transactionRequest = new TransactionRequest
+            {
+                Amount = 100.00M,
+                MerchantAccountId = MerchantAccountIDs.AFT_FIRST_DATA_WALLET_TRANSFER,
+                CreditCard = new TransactionCreditCardRequest
+                {
+                    Number = "4111111111111111",
+                    ExpirationDate = "06/2026",
+                    CVV = "123"
+                },
+                Transfer = new TransferRequest
+                {
+                    Type = "wallet_transfer",
+                    Receiver = new ReceiverRequest
+                    {
+                        FirstName = "Bob",
+                        LastName = "Souza",
+                        MiddleName = "A",
+                        AccountReferenceNumber = "123456789",
+                        AccountReferenceNumberType = "INVALID_ACCOUNT_REFERENCE_NUMBER_TYPE",
+                        Address = new AddressRequest
+                        {
+                            StreetAddress = "1st Main, door 12th",
+                            Locality = "LA",
+                            Region = "CA",
+                            CountryCodeAlpha2 = "US"
+                        }
+                    }
+                }
+            };
+
+            Result<Transaction> result = gateway.Transaction.Sale(transactionRequest);
+            Assert.IsFalse(result.IsSuccess());
+            Assert.AreEqual(ValidationErrorCode.TRANSACTION_TRANSFER_RECEIVER_ACCOUNT_REFERENCE_NUMBER_TYPE_IS_NOT_VALID, result.Errors.ForObject("accountFundingTransaction").OnField("receiver_account_reference_number_type")[0].Code);
         }
     }
 }